requirements.txt
streamlit>=1.35.0
langchain-groq>=0.1.9
langchain-core>=0.2.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
pymupdf4llm>=0.0.17
tavily-python>=0.3.3
streamlit-mic-recorder>=0.0.8
edge-tts>=6.1.12
python-dotenv>=1.0.1
SpeechRecognition>=3.10.4
pydub>=0.25.1
pandas>=2.0.0
matplotlib>=3.7.0
pytesseract>=0.3.10
pdf2image>=1.17.0
plotly>=5.18.0
supabase>=2.0.0
cachetools>=5.3.0
aiohttp>=3.9.0
python-docx>=1.1.0
reportlab>=4.0.0
seaborn>=0.13.0
scikit-learn>=1.3.0
googletrans==4.0.0rc1
streamlit-option-menu>=0.3.6
streamlit-toast>=0.1.0
streamlit-elements>=0.1.0
